create LikePattern operand
